fix: algolia action
